Changed routing for notifications to separate endpoint for my/admin endpoint
